рефакторинг после код ревью
