+++ lib/feature/presentation/ui/news_page/articles/single_article/single_news_article_mobile.dart
@@ -58,7 +58,7 @@ class SingleNewsArticleMobile extends StatelessWidget {
                         enableFeedback: false,
                         icon: const Icon(Icons.arrow_back),
                         iconSize: 24,
-                        onPressed: () => GoRouter.of(context).pop(),
+                        onPressed: () => context.pop(),
                       ),
                     ),
                   ),
